requirements.txt
selenium==3.141.0
chromedriver-binary==2.37.0
requests-oauthlib>=1.2.0
python-quickbooks>=0.9.1
urllib3>=1.24.1
oauthlib>=3.0.1
botocore>=1.12.134
boto3
beautifulsoup4
wheniwork
openpyxl
google-api-python-client
google-auth-httplib2
